requirements.txt
slack_bolt>=1.18.0
openai>=1.35.0
python-dotenv>=1.0.0
